Implement depth only Surveys
Changed the code (readFrom File, densify survey, calcXYZ with
DrillholeCoordBuilder class) so that survey data is read in from a file
in the form of Holeid, depth dip azi (cf HoleID, from, To, dip azi).
This conforms to the more common useage and recording of this type of
data.  Also cleared out some outdated comments/commented out code that
is well out of date. Tested and appears to work as intended. Fixes #8